move some paex symbols to shared
